Nova formatação e Remoção de erros

+++ src/Interfaces/Controlador/ControladorCompraInterface.pas
@@ -11,7 +11,7 @@ type
     ['{7DAF8176-A69A-4386-9A27-A7694AF26C81}']
     procedure ExibirProdutos;
 
-    function AdicionarProduto(Produto: TProduto; Quantidade: integer; PrecoSubtotal: Double): Boolean;
+    function AdicionarProduto(Produto: TProduto; Quantidade: Double; PrecoSubtotal: Double): Boolean;
     function ObterProdutos: TArray<TProdutoQuantidade>;
   end;
